Small LocalSearch refactor
