+++ core/src/main/java/es/urjc/etsii/grafo/solver/improve/IteratedImprover.java
@@ -19,10 +19,10 @@ public abstract class IteratedImprover<S extends Solution<I>,I extends Instance>
     public S improve(S s) {
         int rounds = 0;
         while (!MorkLifecycle.stop() && iteration(s)){
-            log.fine(String.format("Executed iteration %s for LS %s", rounds, this.getClass().getSimpleName()));
+            log.fine(String.format("Executing iteration %s for %s", rounds, this.getClass().getSimpleName()));
             rounds++;
         }
-        log.fine(String.format("LS: %s executed %s iterations", this.getClass().getSimpleName(), rounds));
+        log.fine(String.format("Improvement ended. %s executed %s iterations.", this.getClass().getSimpleName(), rounds));
         return s;
     }
