+++ core/src/main/java/es/urjc/etsii/grafo/solver/improve/ls/LocalSearch.java
@@ -1,12 +1,15 @@
-package es.urjc.etsii.grafo.solver.improve;
+package es.urjc.etsii.grafo.solver.improve.ls;
 
 import es.urjc.etsii.grafo.io.Instance;
 import es.urjc.etsii.grafo.solution.Move;
 import es.urjc.etsii.grafo.solution.MoveComparator;
 import es.urjc.etsii.grafo.solution.neighborhood.Neighborhood;
 import es.urjc.etsii.grafo.solution.Solution;
+import es.urjc.etsii.grafo.solver.improve.DefaultMoveComparator;
+import es.urjc.etsii.grafo.solver.improve.IteratedImprover;
 
 import java.util.Arrays;
+import java.util.Optional;
 
 public abstract class LocalSearch<M extends Move<S, I>, S extends Solution<I>, I extends Instance> extends IteratedImprover<S, I> {
     protected final Neighborhood<M, S, I>[] providers;
@@ -47,8 +50,19 @@ public abstract class LocalSearch<M extends Move<S, I>, S extends Solution<I>, I
         this(new DefaultMoveComparator<>(maximizing), "", ps);
     }
 
+    @Override
+    public boolean iteration(S s) {
+        // Buscar el move a ejecutar
+        var move = getMove(s);
 
-    public abstract boolean iteration(S s);
+        if (move.isEmpty()) {
+            return false; // No existen movimientos válidos, finalizar
+        }
+
+        // Ejecutamos el move y pedimos otra iteracion
+        move.get().execute();
+        return true;
+    }
 
   
     @Override
@@ -69,6 +83,6 @@ public abstract class LocalSearch<M extends Move<S, I>, S extends Solution<I>, I
      * @param s Solution
      * @return Proposed move
      */
-    protected abstract M getMove(S s);
+    protected abstract Optional<M> getMove(S s);
 
 }
