+++ core/src/main/java/es/urjc/etsii/grafo/solver/improve/ls/LocalSearchFirstImprovement.java
@@ -1,10 +1,12 @@
-package es.urjc.etsii.grafo.solver.improve;
+package es.urjc.etsii.grafo.solver.improve.ls;
 
 import es.urjc.etsii.grafo.io.Instance;
 import es.urjc.etsii.grafo.solution.Move;
 import es.urjc.etsii.grafo.solution.MoveComparator;
 import es.urjc.etsii.grafo.solution.Solution;
 import es.urjc.etsii.grafo.solution.neighborhood.Neighborhood;
+
+import java.util.Optional;
 
 public class LocalSearchFirstImprovement<M extends Move<S, I>, S extends Solution<I>, I extends Instance> extends LocalSearch<M, S, I> {
 
@@ -24,23 +26,8 @@ public class LocalSearchFirstImprovement<M extends Move<S, I>, S extends Solutio
         super(maximizing, lsName, ps);
     }
 
-
     @Override
-    public boolean iteration(S s) {
-        // Buscar el move a ejecutar
-        var move = getMove(s);
-
-        if (move == null) {
-            return false; // No existen movimientos válidos, finalizar
-        }
-        assert move.improves();
-        // Ejecutamos el move y pedimos otra iteracion
-        move.execute();
-        return true;
-    }
-
-    @Override
-    protected M getMove(S s) {
+    protected Optional<M> getMove(S s) {
         M move = null;
         for (var provider : providers) {
             var optionalMove = provider.stream(s).filter(Move::isValid).filter(Move::improves).findAny();
@@ -52,7 +39,7 @@ public class LocalSearchFirstImprovement<M extends Move<S, I>, S extends Solutio
                 move = comparator.getBest(move, _move);
             }
         }
-        return move;
+        return move != null && move.improves() ? Optional.of(move) : Optional.empty();
     }
 
 }
